Updated error messages for methods

+++ src/main/java/se/iths/rest/StudentRest.java
@@ -1,5 +1,7 @@
 package se.iths.rest;
 
+import se.iths.exception.BadRequestException;
+import se.iths.exception.NotFoundException;
 import se.iths.entity.Student;
 import se.iths.service.StudentService;
 
@@ -8,8 +10,6 @@ import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
 
 @Path("students")
@@ -24,30 +24,41 @@ public class StudentRest {
     @Path("")
     @POST
     public Response createStudent(Student student) {
+
+        if (student.getFirstName().isEmpty() || student.getLastName().isEmpty() || student.getEmail().isEmpty()) {
+            throw new BadRequestException("You have to fill in firstname, lastname and email");
+        }
         studentService.createStudent(student);
+
         return Response.ok(student).status(Response.Status.CREATED).build();
     }
 
 
     //GET one
-
     @Path("{id}")
     @GET
     public Response getStudent(@PathParam("id") Long id) {
-        Student foundStudent = studentService.findStudentById(id);
-        if (foundStudent == null) {
-            throw new WebApplicationException(Response.status(Response.Status.NOT_FOUND)
-                    .entity("Student with id " + id + " not found.").type(MediaType.APPLICATION_JSON).build());
-        }
-            return Response.ok(foundStudent).status(Response.Status.FOUND).build();
 
+        Student foundStudent = studentService.findStudentById(id);
+
+        if (foundStudent == null) {
+            throw new NotFoundException("Student with id " + id + " not found.");
+        }
+
+        return Response.ok(foundStudent).status(Response.Status.FOUND).build();
     }
 
     //GET all
     @Path("")
     @GET
     public Response getAllStudents() {
+
         List<Student> allStudents = studentService.getStudents();
+
+        if (allStudents.isEmpty()) {
+            throw new NotFoundException("No students enrolled.");
+        }
+
         return Response.ok(allStudents).status(Response.Status.FOUND).build();
     }
 
@@ -57,7 +68,7 @@ public class StudentRest {
     @GET
     public Response getStudentByLastName(@QueryParam("lastname") String lastname) {
 
-        List<Student> students = studentService.findStudentByLastMame(lastname);
+        List<Student> students = studentService.findStudentByLastName(lastname);
 
         List<Student> foundStudents = new ArrayList<>();
 
@@ -66,6 +77,9 @@ public class StudentRest {
                 foundStudents.add(s);
             }
 
+        if (foundStudents.isEmpty()) {
+            throw new NotFoundException("No student with lastname " + lastname + " is found.");
+        }
         return Response.ok(foundStudents).status(Response.Status.FOUND).build();
     }
 
@@ -74,6 +88,13 @@ public class StudentRest {
     @Path("")
     @PUT
     public Response updateStudent(Student student) {
+
+        if (student.getId() == null)
+            throw new BadRequestException("You have to enter student ID number to update contact information.");
+        if (student.getFirstName().isEmpty() || student.getLastName().isEmpty() || student.getEmail().isEmpty()) {
+            throw new BadRequestException("You have to fill in firstname, lastname and email. ");
+        }
+
         studentService.updateStudent(student);
         return Response.ok(student).build();
     }
@@ -83,6 +104,13 @@ public class StudentRest {
     @Path("{id}")
     @DELETE
     public Response deleteStudent(@PathParam("id") Long id) {
+
+        Student foundStudent = studentService.findStudentById(id);
+
+        if (foundStudent == null) {
+            throw new NotFoundException("Student with id " + id + " not found.");
+        }
+
         studentService.deleteStudent(id);
         return Response.ok().build();
     }
